requirements.txt
numpy==1.22.3
pyserial==3.5
scikit_learn==1.1.2
tqdm==4.64.0
